Do what I said in 8b23d5be319ed4bac5d855c127dcfe7c6d0dc32c
Don't mark the template class TableDataHandler for DLL export/import
either. It seems to break the MSVC build, and dropping the decoration
fixed the problem.

Odd that I didn't notice before now, or is it really over a month
since I managed to build master last? That wouldn't be too surprising,
I guess. But other people do claim the have built master with MSVC
relatively recently. Or is it just that this templace class has only
recently been used more than earlier? Or does the fact that I am still
attempting a --enable-dbgutil build affect this? The world is full of
little mysteries.