Improve performance of large Excel documents wrt cell style import.
This change improves import performance of Excel documents based on
the following changes.

1) Pool styles only once per call.

This removes SfxItemPoolCache instantiation and destruction which
is quite expensive.  This alone cuts import almost by half with
large documents with lots of styles applied.

2) Skip removal of direct formats when setting style.

The old code was iterating through all attribute regions in order
to try to remove overlapping format attributes when applying a
style.  But that's strictly not necessary (and bad for performance)
during import.  This also removes the need to re-apply the direct
formats right afterward.

3) Avoid creating a duplicate ScPatternAttr instance per XF.

There is simply no need to create a clone of the style only to put
it into the styles pool and get discarded.  This saves additional
few seconds.