Bypass building the iOS cppu_unittester_all for now
We can't build it here at this point during the build after all. We
use static linking on iOS, and we would need to link with -lgcc3_uno
from bridges which has not been built yet. It seems that any
meaningful unit tests for iOS actually need to be built in a
"subsequent" stage.