Text Boundaries: draw small inner corners for the columns areas
Drawing small inner corners for the columns helps showing the columns
without drawing a full rectangle and it is still compatible with the
page and sections text boundaries.

The whole subsidiary lines painting has been changed to call the method
on all nested SwLayoutFrm. In order not to draw anything, simply
override the method and do nothing in it: (almost) every object takes
care of its boundaries.