Let's avoid modifying the selection data during search and replace.
The previous code was modifying the selection data (ScMarkData) in
a not-so-obvious fashion during the search and/or replace.  Let's only
modify selection in the view code to avoid surprises.  The document
model shouldn't be modifying the view model.

Those methods that were taking a reference to ScMarkData now take a
const reference instead.