Don't generate config.status twice
The current configure script generates ooo.lst for the purpose of extracting
the filename for the seamonkey tarball. This is unnecessary because ooo.lst.in
contains the exact same filename. This patch moves the generation of ooo.lst
next to the other generated files. This move uncovered that the code to set
PREFIXDIR was depending on the AC_OUTPUT code to set prefix and exec_prefix
to /usr/local or the value of --prefix  rather than NONE. Because of this I
added the prefix setting code from AC_OUTPUT to the code that sets PREFIXDIR.