gb_Library_set_componentfile: depend from workdir
The dependecy from the outdir lib to the component file is racy, because
we preserve timestamps when delivering: if the component file and the
workdir library are built in parallel, and the component file ends up
being newer, then the library will always be delivered because its
timestamp is always older.  Work around this by having the workdir
library depend on the component file.