Work around a problem creating symlinks in ooinstall.
It can happen that the ure-link symlink shall be created while its
parent directory has not yet been created.  The logic in the Perl code
is so confusing and beyond fixing, that an easy hack out was to flag
the respective directory as always-CREATE, which kind of works around
the problem.