Resolves: fdo#32665 handle that FreeSerif lacks some glyphs in bold/italic
FreeSerif lacks glyphs in bold/italic variants that it has in the normal one. A
lot of our glyph fallback infrastructure, especially the caches don't expect
that a normal variant of a font with extra emboldening or extra font skew can
be a fallback for a bold/italic variant of itself which exists, but doesn't
have the missing glyphs that the normal one does.

We really need to improve our glyph/font caching, but we can get 90% of the
way there by excluding such cases from the caches.