Unwind and refactor some of the Grey Palette stuff
We get some assertions in the smoketest with --enable-debug, looking
closer I think to use a single bIsGrey flag we...

a) would need to check after every modification of a Palette that mbIsGrey is
still valid and update it accordingly
b) There are a lot of routes outside of direct assignation to the stock Grey
Palettes that can result in Grey palettes in e.g. the pngreader, they would
all need updating :-(

So, how about just refactoring the original HasGreyPalette and IsGreyPalette.