We want Android shared objects to have names ending in .so
The silly apkbuilder tool doesn't add extra native libs to an .apk
package unless their name ends with .so. It just silently ignores
them. So, force libtool to avoid versioning suffix for Android.

Yeah, unclear whether apkbuilder is what we will end up using to
construct the APKs of actual LO code using Android apps. But at this
stage when just trying to build a unit testing app, it seems to be the
simplest way to get the .apk properly signed and all to just use a
normal Ant project, and let Ant run apkbuilder.