Make the toolbars not popping-up randomly
Undocked, context sensitive toolbars often tend to overlay important parts of
a document.
With this patch these toolbars are docked to the bottom to prevent the document from
being moved each time the toolbar is displayed.
Furthermore the findbar setting was changed to make sure that it stays
on top of the other toolbars at the bottom.