Look for libboostrap.uno.so and not bootstrap.uno.so on Android
Not sure if we need to be more generic here, and always add a "lib"
prefix if not present. Or do the changes to the makefilery take care
of it, so that for other UNO components than bootstrap, the name as
stored wherever it is stored does contain the "lib" prefix we are
forces to use on Android?