i#102333: SwDoc::ReplaceRangeImpl:
The fix for this (d0b32daf6062146b050d6d336253bc352e044ae7) was not
quite right; there are 2 different join modes, and moving the cursor
backward is required in one of them, but not the other.

This can be seen when doing replace all ^$ with several consecutive
empty lines.