Fix for bug fdo39748, Easy-hack Cleanup extension list.
This patch introduces 3 new check box in the extension manager GUI to allow selection of extension type to display: bundled, shared or user. Dialog is automatically updated upon toggling checkbox.

On toggling each checkbox, the extension list is recreated from scratch and packages are added to the list depending on the checkmark state. Initial state is all checks marked.