SwPosition: handle un-registered SwIndex:
For frames anchored at paragraph, SwFmtAnchor::SetAnchor() creates
an anchor SwPosition that points at the SwTxtNode but with a SwIndex
that is not registered at the SwTxtNode.
SwXFrame::getAnchor() then creates a bookmark from that, which asserts
in the SwIndex comparison operators.

Work around this by adding ugly special handling for the case of 2
positions having their SwNodeIndex point to the same node but one of
them not having its SwIndex at the node.

It seems it is best to do this in SwPosition operators, as the SwIndex
operators lack the context to know if this inconsistency is allowed,
and it is unknown whether the bookmarks sort code is the only path that
triggers this assert.